+++ token_throttle/_limiter_backends/_redis/_sync_backend.py
@@ -395,7 +395,7 @@ class SyncRedisBackend(SyncRateLimiterBackend):
             )
 
             # Apply refund amounts to current capacity
-            updated_capacities_: dict[tuple[str, float], float] = dict(
+            updated_capacities_: dict[tuple[str, int], float] = dict(
                 prerefund_capacities,
             )
             for (
@@ -428,7 +428,7 @@ class SyncRedisBackend(SyncRateLimiterBackend):
                         ),
                         bucket.max_capacity,
                     )
-                updated_capacities = frozendict(updated_capacities_)
+            updated_capacities = frozendict(updated_capacities_)
 
             # Always update capacities in Redis with the current time
             self._set_capacities_unsafe(
